requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bitarray==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.7
ckzg==1.0.2
cytoolz==1.1.0
eth-account==0.11.3
eth-hash==0.8.0
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==5.2.0
frozenlist==1.8.0
h11==0.16.0
hexbytes==0.3.1
httpcore==1.0.9
httpx==0.25.2
idna==3.11
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lru-dict==1.2.0
multidict==6.7.1
parsimonious==0.10.0
pillow==12.2.0
propcache==0.4.1
protobuf==7.34.1
py-ecc==8.0.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-telegram-bot==20.7
pyunormalize==17.0.0
referencing==0.37.0
regex==2026.4.4
reportlab==4.4.10
requests==2.33.1
rlp==4.1.0
rpds-py==0.30.0
sniffio==1.3.1
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
web3==6.15.1
websockets==16.0
yarl==1.23.0
